mcp-demo_python-sdk/requirements.txt
mcp-server
mcp-client
ollama
